created specific Gallery component and imported that in Explore.tsx
