+++ frontend/src/pages/Explore/Explore.tsx
@@ -2,8 +2,9 @@ import React, { useContext, useEffect, useState } from 'react';
 import { LoginContext } from '../../Components/LoginManager/LoginManager';
 import { getSightings } from '../../api/backendClient';
 import "./Explore.scss"
+import Gallery from '../../Components/Gallery/Gallery';
 
-interface SightingType {
+export interface SightingType {
   id: number
   username: string;
   speciesName: string;
@@ -14,48 +15,11 @@ interface SightingType {
   dateTime: Date;
 }
 
-const formatDate = (dateTime: Date) => {
-  return new Date(dateTime).toLocaleString("en-GB", { timeZone: "UTC" })
-}
-
 function Explore(): JSX.Element {
-  const [sightings, setSightings] = useState<SightingType[]>([]);
-  const loginContext = useContext(LoginContext);
-  const jwt = loginContext.jwt;
-
-  useEffect(() => {
-    async function fetchSightings() {
-      try {
-
-        const response = await getSightings(jwt);
-        if (!response.ok) {
-          throw new Error('Network response failed');
-        }
-        const result = await response.json();
-        setSightings(result.sightings);
-      }
-      catch (error) {
-        console.error('Error fetching sightings:', error);
-      }
-    }
-
-    fetchSightings();
-  }, [jwt]);
 
   return (
     <>
-      <h1>Explore</h1>
-      {sightings.map(sighting => (
-        <>
-        <p>{sighting.username}</p>
-        <p>{sighting.speciesName}</p>
-        <img src={sighting.photoUrl} className="thumbnail-explore-gallery"/>
-        <p>Latitude: {sighting.latitude}</p>
-        <p>Longitude: {sighting.longitude}</p>
-        {sighting.description && <p>Description: {sighting.description}</p> }  
-        <p>Date: {formatDate(sighting.dateTime)}</p>
-        </>
-      ))}
+      <Gallery />
     </>
   )
